Change launch method and update README.
The method used before to select the controller from a launchfile argument made
it impossible to provide arguments to the plugins in the Gazebo way. This has
been reworked into controller-centric xacro system, in which each controller
has its own xacro file which needs to load the slider description.

The README has been updated to reflect this changes and to include a
description of the AdmittanceControllerBase introduced in the last commit.